Handle field privacy when a profile field group is disassociated from a hub.
If a profile field group is disassociated from all hubs, don’t leave
the profile field visibility setting as “groupadmins” which has no
meaning outside of profile groups that are associated with hubs. So we
change the default setting for fields in that group if necessary, and
also for users who have saved that field with the affected visibility
setting.