Add specific nonce to extra form and save routine.
In some cases, group membership requests are sent from the registration
page, not the group membership request page, as was expected in
`process_group_requests_profile_form()`. To avoid the problem of
`check_admin_referer()` failing (and calling `die()`!), set a specific
nonce for these fields and check it at save, rather than relying on the
BP request form.